handle updating unclaimed in `BridgeProvider`
